create new folder - repositories - to store files that interact with DB, extract model from repository and create dataModel file
