+++ api/routes/products.js
@@ -3,7 +3,7 @@ const router = express.Router();
 
 const ProductsController = require('../controllers/productsController');
 
-router.get('/', ProductsController.get_all);
+router.get('/', ProductsController.getAll);
 router.get('/:productId', ProductsController.getById);
 router.post('/', ProductsController.create);
